同步改动

+++ Scripts/AutoBuild_Function.sh
@@ -12,15 +12,14 @@ GET_TARGET_INFO() {
 	[[ -z "${Author}" ]] && {
 		Author="$(echo "${User_Repo}" | egrep -o "[a-zA-Z0-9]+" | awk 'NR==4')"
 	}
-	Openwrt_Repo="$(grep "https://github.com/[a-zA-Z0-9]" ${Home}/.git/config | cut -c8-100 | sed 's/^[ \t]*//g')"
 	Openwrt_Author="$(echo "${Openwrt_Repo}" | egrep -o "[a-zA-Z0-9]+" | awk 'NR==4')"
-	Current_Branch="$(git branch | sed 's/* //g' | sed 's/^[ \t]*//g')"
+	Current_Branch="$(GET_BRANCH)"
 	In_Firmware_Info=package/base-files/files/etc/openwrt_info
 	[[ ! ${Current_Branch} == master ]] && {
 		Current_Branch="$(echo ${Current_Branch} | egrep -o "[0-9]+.[0-9]+")"
 		Openwrt_Version_="R${Current_Branch}-"
 	} || {
-		Openwrt_Version_="R18.06-"
+		Openwrt_Version_="R$(date +%y.%m)-"
 	}
 	case ${Openwrt_Author} in
 	coolsnowwolf)
@@ -67,7 +66,11 @@ GET_TARGET_INFO() {
 	esac
 	TARGET_SUBTARGET="$(awk -F '[="]+' '/TARGET_SUBTARGET/{print $2}' .config)"
 
-	echo "Firmware_Type=${Firmware_Type}" > ${Home}/TARGET_INFO
+	echo "Author=${Author}" > ${Home}/TARGET_INFO
+	echo "Default_Device=${Default_Device}" >> ${Home}/TARGET_INFO
+	echo "INCLUDE_Obsolete_PKG_Compatible=${INCLUDE_Obsolete_PKG_Compatible}" >> ${Home}/TARGET_INFO
+	echo "Upload_VM_Firmware=${Upload_VM_Firmware}" >> ${Home}/TARGET_INFO
+	echo "Firmware_Type=${Firmware_Type}" >> ${Home}/TARGET_INFO
 	echo "TARGET_PROFILE=${TARGET_PROFILE}" >> ${Home}/TARGET_INFO
 	echo "Openwrt_Version=${Openwrt_Version}" >> ${Home}/TARGET_INFO
 	echo "Openwrt_Author=${Openwrt_Author}" >> ${Home}/TARGET_INFO
@@ -75,6 +78,7 @@ GET_TARGET_INFO() {
 	echo "TARGET_SUBTARGET=${TARGET_SUBTARGET}" >> ${Home}/TARGET_INFO
 	echo "Home=${Home}" >> ${Home}/TARGET_INFO
 	echo "Current_Branch=${Current_Branch}" >> ${Home}/TARGET_INFO
+	echo "Upload_VM_Firmware=${Upload_VM_Firmware}" >> ${Home}/TARGET_INFO
 	
 	echo "Github=${User_Repo}" > ${In_Firmware_Info}
 	echo "CURRENT_Version=${Openwrt_Version}" >> ${In_Firmware_Info}
@@ -105,24 +109,24 @@ Firmware-Diy_Base() {
 		Replace_File Scripts/AutoBuild_Tools.sh package/base-files/files/bin
 	}
 	[[ "${INCLUDE_AutoUpdate}" == true ]] && {
-		AddPackage git lean luci-app-autoupdate https://github.com/Hyy2001X main
+		AddPackage git lean luci-app-autoupdate Hyy2001X main
 		Replace_File Scripts/AutoUpdate.sh package/base-files/files/bin
 	}
 	[[ "${INCLUDE_Theme_Argon}" == true ]] && {
 		case ${Openwrt_Author} in
 		coolsnowwolf)
-			AddPackage git lean luci-theme-argon https://github.com/jerrykuku 18.06
+			AddPackage git lean luci-theme-argon jerrykuku 18.06
 		;;
 		*)
 			case ${Current_Branch} in
 			19.07)
-				AddPackage git other luci-theme-argon https://github.com/jerrykuku v2.2.5
+				AddPackage git other luci-theme-argon jerrykuku v2.2.5
 			;;
 			21.02)
-				AddPackage git other luci-theme-argon https://github.com/jerrykuku
+				AddPackage git other luci-theme-argon jerrykuku master
 			;;
 			18.06)
-				AddPackage git other luci-theme-argon https://github.com/jerrykuku 18.06
+				AddPackage git other luci-theme-argon jerrykuku 18.06
 			;;
 			*)
 				TIME "[ERROR] Unknown Openwrt branch: [${Current_Branch}] !"
@@ -135,8 +139,6 @@ Firmware-Diy_Base() {
 	if [[ -n "${Defined_IP_Address}" ]];then
 		TIME "Using defined IP Address [${Defined_IP_Address}] ..."
 		New_IP_Address="${Defined_IP_Address}"
-	else
-		TIME "Using default IP Address [${New_IP_Address}] ..."
 	fi
 	[[ -n "${New_IP_Address}" ]] && [[ "${New_IP_Address}" != false ]] && {
 		if [[ "${New_IP_Address}" =~ ^[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}$ ]];then
@@ -144,8 +146,8 @@ Firmware-Diy_Base() {
 			if [[ ! "${New_IP_Address}" == "${Old_IP_Address}" ]];then
 				TIME "Setting default IP Address to ${New_IP_Address} ..."
 				sed -i "s/${Old_IP_Address}/${New_IP_Address}/g" package/base-files/files/bin/config_generate
-				a=$(echo ${Old_IP_Address} | egrep -o "[0-9]+.[0-9]+.")
-				b=$(echo ${New_IP_Address} | egrep -o "[0-9]+.[0-9]+.")
+				a=$(echo ${Old_IP_Address} | egrep -o "[0-9]+.[0-9]+." | awk 'NR==1')
+				b=$(echo ${New_IP_Address} | egrep -o "[0-9]+.[0-9]+." | awk 'NR==1')
 				c="$(egrep -o ")).[0-9]+" package/base-files/files/bin/config_generate)"
 				d=")).$(echo ${New_IP_Address} | egrep -o "[0-9]+" | awk 'END {print}')"
 				sed -i "s/${a}/${b}/g" package/base-files/files/bin/config_generate
@@ -164,14 +166,13 @@ Firmware-Diy_Base() {
 	coolsnowwolf)
 		Replace_File CustomFiles/Depends/coremark_lede.sh package/lean/coremark coremark.sh
 		Replace_File CustomFiles/Depends/cpuinfo_x86 package/lean/autocore/files/x86/sbin cpuinfo
-		AddPackage git other helloworld https://github.com/fw876 master
+		AddPackage git other helloworld fw876 master
 		sed -i 's/143/143,8080/' $(PKG_Finder d package luci-app-ssr-plus)/root/etc/init.d/shadowsocksr
 		sed -i "s?iptables?#iptables?g" ${Version_File}
 		sed -i "s?${Old_Version}?${Old_Version} Compiled by ${Author} [${Display_Date}]?g" ${Version_File}
 		[[ "${INCLUDE_DRM_I915}" == true ]] && Replace_File CustomFiles/Depends/i915-5.4 target/linux/x86 config-5.4
 	;;
 	immortalwrt)
-		sed -i 's/143/143,8080/' $(PKG_Finder d package luci-app-ssr-plus)/root/etc/init.d/shadowsocksr
 		Replace_File CustomFiles/Depends/coremark_ImmortalWrt.sh package/base-files/files/etc coremark.sh
 		Replace_File CustomFiles/Depends/ImmortalWrt package/base-files/files/etc openwrt_release
 		Replace_File CustomFiles/Depends/cpuinfo_x86 package/lean/autocore/files/x86/sbin cpuinfo
@@ -191,29 +192,46 @@ Firmware-Diy_Base() {
 		sed -i "s?Openwrt?Openwrt ${Openwrt_Version} / AutoUpdate ${AutoUpdate_Version}?g" package/base-files/files/etc/banner
 	;;
 	esac
-	[[ "${INCLUDE_Obsolete_PKG_Compatible}" == true ]] && {
+	TIME "[Firmware-Diy_Base] All done !"
+}
+
+Other_Scripts() {
+	[ -f TARGET_INFO ] && . TARGET_INFO || {
+		TIME "[ERROR] No variable_define file detected!"
+		exit 1
+	}
+	case ${INCLUDE_Obsolete_PKG_Compatible} in
+	19.07)
+		Current_Branch=19.07
+		Force_mode=1
+		INCLUDE_Obsolete_PKG_Compatible=true
+	;;
+	21.02)
+		Current_Branch=21.02
+		Force_mode=1
+		INCLUDE_Obsolete_PKG_Compatible=true
+	;;
+	esac
+	if [[ "${INCLUDE_Obsolete_PKG_Compatible}" == true ]];then
 		TIME "Start to run Obsolete_Package_Compatible Scripts ..."
-		[[ ${Openwrt_Author} == openwrt ]] && {
+		if [[ ${Openwrt_Author} == openwrt ]] || [[ "${Force_mode}" == 1 ]];then
 			case ${Current_Branch} in
 			19.07 | 21.02)
 				Replace_File CustomFiles/Patches/0003-upx-ucl-${Current_Branch}.patch ./
 				cat 0003-upx-ucl-${Current_Branch}.patch | patch -p1 > /dev/null 2>&1
-				AddPackage svn ../feeds/packages/lang golang https://github.com/coolsnowwolf/packages/trunk/lang
+				# AddPackage svn feeds/packages/lang golang coolsnowwolf/packages/trunk/lang
 				TIME "Start to convert zh-cn translation files to zh_Hans ..."
 				Replace_File Scripts/Convert_Translation.sh package
-				cd ./package
-				bash ./Convert_Translation.sh
-				cd ..
+				cd ./package && bash ./Convert_Translation.sh && cd ..
 			;;
 			*)
-				TIME "[ERROR] Current branch: [${Current_Branch}] is not supported !"
+				TIME "Current branch: [${Current_Branch}] is not supported,skip..."
 			;;
 			esac
-		} || {
-			TIME "[ERROR] Current source: [${Openwrt_Author}] is not supported !"
-		}
-	}
-	TIME "[Firmware-Diy_Base] All done !"
+		else
+			TIME "Current source: [${Openwrt_Author}] is not supported,skip..."
+		fi
+	fi
 }
 
 PS_Firmware() {
@@ -245,10 +263,12 @@ PS_Firmware() {
 	;;
 	esac
 	Firmware_Path="bin/targets/${TARGET_BOARD}/${TARGET_SUBTARGET}"
+	rm -rf ${Firmware_Path}/packages
 	Mkdir bin/Firmware
 	case "${TARGET_PROFILE}" in
 	x86_64)
 		cd ${Firmware_Path}
+		Default_Firmware="${_Firmware}-${TARGET_BOARD}-${TARGET_SUBTARGET}"
 		Legacy_Firmware="${_Firmware}-${TARGET_BOARD}-${TARGET_SUBTARGET}-${_Legacy_Firmware}.${Firmware_Type}"
 		EFI_Firmware="${_Firmware}-${TARGET_BOARD}-${TARGET_SUBTARGET}-${_EFI_Firmware}.${Firmware_Type}"
 		AutoBuild_Firmware="AutoBuild-${TARGET_PROFILE}-${Openwrt_Version}"
@@ -258,23 +278,31 @@ PS_Firmware() {
 		if [ -f "${Legacy_Firmware}" ];then
 			_MD5=$(md5sum ${Legacy_Firmware} | cut -d ' ' -f1)
 			_SHA256=$(sha256sum ${Legacy_Firmware} | cut -d ' ' -f1)
-			touch ${Home}/bin/Firmware/${AutoBuild_Firmware}.detail
-			echo -e "\nMD5:${_MD5}\nSHA256:${_SHA256}" > ${Home}/bin/Firmware/${AutoBuild_Firmware}-Legacy.detail
-			mv -f ${Legacy_Firmware} ${Home}/bin/Firmware/${AutoBuild_Firmware}-Legacy.${Firmware_Type}
+			echo -e "MD5:${_MD5}\nSHA256:${_SHA256}" > ${Home}/bin/Firmware/${AutoBuild_Firmware}-Legacy.detail
+			cp ${Legacy_Firmware} ${AutoBuild_Firmware}-Legacy.${Firmware_Type}
 			TIME "Legacy Firmware is detected !"
-		else
-			TIME "[ERROR] Legacy Firmware is not detected !"
 		fi
 		if [ -f "${EFI_Firmware}" ];then
 			_MD5=$(md5sum ${EFI_Firmware} | cut -d ' ' -f1)
 			_SHA256=$(sha256sum ${EFI_Firmware} | cut -d ' ' -f1)
-			touch ${Home}/bin/Firmware/${AutoBuild_Firmware}-UEFI.detail
 			echo -e "\nMD5:${_MD5}\nSHA256:${_SHA256}" > ${Home}/bin/Firmware/${AutoBuild_Firmware}-UEFI.detail
-			cp ${EFI_Firmware} ${Home}/bin/Firmware/${AutoBuild_Firmware}-UEFI.${Firmware_Type}
+			cp ${EFI_Firmware} ${AutoBuild_Firmware}-UEFI.${Firmware_Type}
 			TIME "UEFI Firmware is detected !"
-		else
-			TIME "[ERROR] UEFI Firmware is not detected !"
 		fi
+		if [[ "${Upload_VM_Firmware}" == true ]];then
+			TIME "Converting vmdk firmware to vhdx ..."
+			[ -f "${Default_Firmware}-${_Legacy_Firmware}.vmdk" ] && {
+				qemu-img convert -O vhdx ${Default_Firmware}-${_Legacy_Firmware}.vmdk ${AutoBuild_Firmware}-Legacy.vhdx
+				cp ${Default_Firmware}-${_Legacy_Firmware}.vmdk ${AutoBuild_Firmware}-Legacy.vmdk
+				cp ${Default_Firmware}-${_Legacy_Firmware}.vdi ${AutoBuild_Firmware}-Legacy.vdi
+			}
+			[ -f "${Default_Firmware}-${_UEFI_Firmware}.vmdk" ] && {
+				qemu-img convert -O vhdx ${Default_Firmware}-${_Legacy_Firmware}.vmdk ${AutoBuild_Firmware}-UEFI.vhdx
+				cp ${Default_Firmware}-${_UEFI_Firmware}.vmdk ${AutoBuild_Firmware}-UEFI.vmdk
+				cp ${Default_Firmware}-${_Legacy_Firmware}.vdi ${AutoBuild_Firmware}-Legacy.vdi
+			}
+		fi
+		mv -f AutoBuild-* ${Home}/bin/Firmware
 	;;
 	*)
 		cd ${Home}
@@ -296,6 +324,14 @@ PS_Firmware() {
 	esac
 	cd ${Home}
 	echo "[$(date "+%H:%M:%S")] Actions Avaliable: $(df -h | grep "/dev/root" | awk '{printf $4}')"
+}
+
+GET_BRANCH() {
+    local Folder="$(pwd)"
+    [ -n "$1" ] && Folder="$1"
+    git -C "${Folder}" rev-parse --abbrev-ref HEAD | grep -v HEAD || \
+    git -C "${Folder}" describe --exact-match HEAD || \
+    git -C "${Folder}" rev-parse HEAD
 }
 
 TIME() {
@@ -344,7 +380,7 @@ Auto_AddPackage_mod() {
 	}
 	_FILENAME=${1}
 	echo "" >> ${_FILENAME}
-	[ -f "${_FILENAME}" ] && {
+	[ -f "${_FILENAME}" ] && [ -s "${_FILENAME}" ] && {
 		TIME "Loading Custom Packages list: [${_FILENAME}]..."
 		cat ${_FILENAME} | sed '/^$/d' | while read X
 		do
@@ -365,7 +401,7 @@ AddPackage() {
 		PKG_PROTO=${1}
 		PKG_DIR=${2}
 		PKG_NAME=${3}
-		REPO_URL=${4}
+		REPO_URL="https://github.com/${4}"
 		REPO_BRANCH=${5}
 	;;
 	*)
